+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Next
+- Added `keyword-spacing` to enforce consistent spacing before and after keywords
 - Added `space-before-blocks` to enforce consistent spacing before blocks
 - Added `comma-spacing` to enforce consistent spacing before and after commas
 - Added `indent`/`react/jsx-indent` to enforce consistent indentation